Use mutex rather than critical sections.
For TYPE_LOCK, replace critical sections with a simple mutex.  A number
of changes were needed to avoid deadlocking on reentrancy (trying to
re-acquire the mutex). Split _PyType_LookupRefAndVersion() into a locked
and unlocked version.  Remove atomic operations where they are not
required. Remove some cases of TYPE_LOCK being held that is not
required.