gh-115028: Add support for relocatable installations using $ORIGIN in RPATH
This patch adds a new configure option --with-relative-rpath that, when
enabled, adds $ORIGIN-based RPATHs to binaries when building Python on Linux.
This makes Python installations relocatable so they can be moved to different
directories without breaking dynamic library dependencies.

Three new variables are introduced to set RPATHs for different components:
- PY_RPATH_EXEC: For the Python interpreter binary
- PY_RPATH_LIB: For the shared libpython library
- PY_RPATH_MOD: For Python extension modules

This is a step toward solving the relocatability issues discussed in
packaging forums and helps ena