Make remote_exec accept `bytes` paths
This was documented as working, but in fact we assumed that the path
would always be a unicode string. Factor the version that only accepts a
unicode string into a helper function, and have the caller call
`os.fsdecode` on the input path, which accepts either a byte string or a
unicode string and always returns a unicode string.