Do some addtional locking simplification.
Re-structure type_setattro() into a slot and non-slot case.  For the
slot case, we need to stop-the-world.  For non-slot, we need to hold
TYPE_LOCK.

Remove unneeded locking in type_mro_modified().  This is only called if
the type is new or if the world is stopped.  In both cases, TYPE_LOCK
doesn't need to be held.