gh-129210: Skip second reachability check if no finalizers run
To handle objects potentially resurrected in finalizers it is necessary to
repeat the reachability check on objects to be collected. If there are no
finalizers run this can be skipped: in that case it is not possible for an
object to be resurrected.

In my environment and testing this gives a ~25-30% performance improvement for
the gh-129210 micro-benchmark.