gh-129210: Skip second reachability check in ft GC if no finalizers run
As for the regular GIL garbage collector, skip the second reachability check if
no objects to-be-collected have finalizers, and hence none can be resurrected.

In (so far very limited) testing this gives a ~10% overall performance
improvement to the motivating benchmark's runtime and a ~35% improvement to its
reported time.

The pyperformance gc_traversal and gc_collect benchmarks show no significant
difference and a 6% improvement respectively.