Try to prevent a PermissionError on Windows
If `NamedTemporaryFile.__exit__` tries to delete the connect script
before the remote process has closed it, a `PermissionError` is raised
on Windows. Unfortunately there is no synchronization that we can use to
ensure that the file is closed before we try to delete it. This makes
the race less likely to occur by making the `with` block contain more
code, so that more time passes before we attempt to delete the file.

Co-authored-by: Chris Eibl <138194463+chris-eibl@users.noreply.github.com>