gh-92408: cleanup during error handling for POSIX shared memory
`multiprocessing.shared_memory.SharedMemory` currently tries to unlink a POSIX
shared memory object if it fails to `mmap` it after opening/creating it.

However, since it only catches `OSError`, it fails to do so when opening an
existing zero-length shared memory object, as `mmap.mmap` raises a `ValueError`
when given a 0 size.

Fix this by catching all exceptions with a bare `except`, so the cleanup is run
for all errors.