Do not unlink pre-existing POSIX shared memory on error
`multiprocessing.shared_memory.SharedMemory` currently attempts to `unlink` a
POSIX shared memory object it has opened or created if a subsequent exception
is raised in its constructor.

This is necessary to avoid it leaking if the shared memory object was created
in the constructor, however it seems inappropriate if it is opening a
pre-existing shared memory object.

Fix this by only unlinking the shared memory object if it was created.