Allow zero-length shared memory objects
These are allowed by POSIX and it seems more consistent/cleaner to support them
than forcing the user to treat them as a special case.

Windows does not support creating a zero-length mapping, and neither does
`mmap` on any platform, so skip those operations and assign a zero-length
buffer instead.

TODO: The way the POSIX/Windows code is interleaved inside if statements makes
for difficult to follow and deeply nested logic. It might be better to refactor
this code and keep the platform-specific logic in separate functions or
possibly even implementation classes.