gh-96026: do not unregister without register in shm POSIX error handling
`multiprocessing.shared_memory.SharedMemory` currently tries to unlink a POSIX
shared memory object if it fails to mmap it after opening/creating it in the
constructor. It does this by calling `cleanup()`, which unregisters the shared
memory despite it not having been registered. This will trigger a separate
exception in the resource tracker.

Fix this by calling the low-level unlink directly.