gh-97907: use cancellation error from child if possible
At present when cancelling the result of an :func:`asncio.gather` call, the
last child is used to create the cancellation error. If that was not cancelled
but another child was, its cancellation message and traceback will be lost.

Fix this by using the cancellation error from the first of the children to be
cancelled, if any, falling back to the last child only if none have been.