Remove unnecessary mention of os.fsync() from os.exec*() docs
Before a process is replaced by os.exec*(), userspace buffers usually
need to be flushed. However, it is not necessary to flush kernel buffers
to disk using os.fsync(). Flushing userspace buffer to kernel, and
flushing kernel buffer to disk, are two different things. In addition,
os.fsync() is not applicable to pipes or sockets. This patch removes
this confusing information.